Minor polishing
