+++ src/sort/sort-impl.h
@@ -1,0 +1,17 @@
+#ifndef _SORT_IMPL_H_
+#define _SORT_IMPL_H_
+
+#include <sort.h>
+
+#define min(a,b) ((a)<(b) ? (a) : (b))
+#define max(a,b) ((a)>(b) ? (a) : (b))
+
+double get_scalar(struct array *a,uint i,uint offset,uint usize,
+    gs_dom type);
+void get_extrema(void *extrema_,sort_data data,uint field,struct comm *c);
+int  set_dest(uint *proc,uint np,ulong start,uint size,ulong nelem);
+int  load_balance(struct array *a,size_t size,struct comm *c,
+  struct crystal *cr);
+int  sort_local(sort_data data);
+
+#endif
